# docs/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "CompetencyX API",
        "version": "0.1.0",
        "description": "Adaptive career-roadmap backend API for assessment sessions, role inference, and next-topic recommendations."
    },
    "paths": {
...
        "/api/v1/assessment-sessions/{id}/answers/": {
            "post": {
                "operationId": "submitAssessmentAnswer",
                "summary": "Submit an answer for the current question",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Assessment session UUID.",
                        "required": true
                    }
                ],
                "tags": [
                    "Assessment Sessions"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnswerSubmit"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "tokenAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssessmentSession"
                                },
                                "examples": {
                                    "SubmitAnswerResponse": {
                                        "value": {
                                            "id": "2b39d41d-8de9-4b9b-b2ef-2a278b3f3770",
                                            "status": "in_progress",
                                            "phase": "role_discovery",
                                            "language": "en",
                                            "best_fit_confidence": 0.0,
                                            "preferred_role": {
                                                "id": 1,
                                                "slug": "backend-engineer",
                                                "name": "Backend Engineer",
                                                "description": "Builds APIs and backend services."
                                            },
                                            "best_fit_role": null,
                                            "profile": {
                                                "education_level": "student",
                                                "current_stage": "beginner"
                                            },
                                            "started_at": "2026-04-17T04:00:00Z",
                                            "updated_at": "2026-04-17T04:00:00Z",
                                            "completed_at": null,
                                            "milestones": {
                                                "answered_role_questions": 0,
                                                "answered_core_role_questions": 0,
                                                "answered_tie_break_questions": 0
                                            },
                                            "role_alignment_status": "unknown",
                                            "role_resolution_status": "in_progress",
                                            "guidance_summary": "You want to pursue Backend Engineer. Complete the role-discovery profile to compare fit.",
                                            "current_question": {
                                                "id": 101,
                                                "code": "role-likert-build-working-parts",
                                                "stage": "role",
                                                "question_type": "likert_5",
                                                "prompt": "I enjoy turning an idea into a working technical part.",
                                                "help_text": "",
                                                "role": null,
                                                "topic": null,
                                                "difficulty": 1,
                                                "options": [],
                                                "response_scale": [
                                                    {
                                                        "key": "strongly_agree",
                                                        "label": "Strongly agree",
                                                        "value": 2,
                                                        "display_order": 1
                                                    },
                                                    {
                                                        "key": "agree",
                                                        "label": "Agree",
                                                        "value": 1,
                                                        "display_order": 2
                                                    },
                                                    {
                                                        "key": "neutral",
                                                        "label": "Neutral",
                                                        "value": 0,
                                                        "display_order": 3
                                                    },
                                                    {
                                                        "key": "disagree",
                                                        "label": "Disagree",
                                                        "value": -1,
                                                        "display_order": 4
                                                    },
                                                    {
                                                        "key": "strongly_disagree",
                                                        "label": "Strongly disagree",
                                                        "value": -2,
                                                        "display_order": 5
                                                    }
                                                ]
                                            }
                                        },
                                        "summary": "Submit answer response"
                                    }
                                }
                            }
                        },
                        "description": "Updated assessment session state after a valid answer submission."
                    },
                    "400": {
                        "description": "Validation or assessment flow error, including out-of-order answers."
                    },
                    "404": {
                        "description": "Assessment session was not found."
                    }
                }
            }
        },
        "/api/v1/assessment-sessions/{id}/history/": {
            "get": {
                "operationId": "getAssessmentHistory",
                "summary": "Get completed assessment history",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Assessment session UUID.",
                        "required": true
                    }
                ],
                "tags": [
                    "Assessment Sessions"
                ],
                "security": [
                    {
                        "tokenAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssessmentHistory"
                                },
                                "examples": {
                                    "HistoryResponse": {
                                        "value": {
                                            "id": "2b39d41d-8de9-4b9b-b2ef-2a278b3f3770",
                                            "phase": "recommendation_ready",
                                            "status": "completed",
                                            "answers": [
                                                {
                                                    "id": 401,
                                                    "question_id": 101,
                                                    "question_code": "role-likert-build-working-parts",
                                                    "question_prompt": "I enjoy turning an idea into a working technical part.",
                                                    "question_stage": "role",
                                                    "topic_slug": null,
                                                    "selected_option_id": null,
                                                    "selected_option_key": null,
                                                    "selected_option_label": null,
                                                    "scale_value": 2,
                                                    "response_time_ms": 4200,
                                                    "confidence_indicator": "high",
                                                    "responded_at": "2026-04-17T04:01:00Z"
                                                }
                                            ]
                                        },
                                        "summary": "History response"
                                    }
                                }
                            }
                        },
                        "description": "Completed assessment answers."
                    },
                    "404": {
                        "description": "Assessment session was not found."
                    },
                    "409": {
                        "description": "Assessment history is only available after completion."
                    }
                }
            }
        },
...
        "/api/v1/assessment-sessions/{id}/skill-assessment/catalog/": {
            "get": {
                "operationId": "getAssessmentSkillAssessmentCatalog",
                "summary": "Get skill assessment PSP and SDLC questionnaire catalog for an assessment session",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Assessment session UUID.",
                        "required": true
                    }
                ],
                "tags": [
                    "Assessment Sessions"
                ],
                "security": [
                    {
                        "tokenAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SkillAssessmentCatalog"
                                }
                            }
                        },
                        "description": "Skill assessment question, dimension, scale, and role guidance catalog."
                    },
                    "404": {
                        "description": "Assessment session was not found."
                    }
                }
            }
        },
        "/api/v1/assessment-sessions/{id}/skill-assessment/next-question/": {
            "post": {
                "operationId": "getAssessmentSkillAssessmentNextQuestion",
                "summary": "Get the next adaptive skill assessment question",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Assessment session UUID.",
                        "required": true
                    }
                ],
                "tags": [
                    "Assessment Sessions"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SkillAssessmentNextQuestionRequest"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "tokenAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SkillAssessmentNextQuestionResponse"
                                },
                                "examples": {
                                    "NextQuestionResponse": {
                                        "value": {
                                            "next_question": {
                                                "id": "q-dev",
                                                "prompt": "I can implement features using clear design and coding practices.",
                                                "dimension_key": "development"
                                            }
                                        },
                                        "summary": "Next question response"
                                    }
                                }
                            }
                        },
                        "description": "Next unanswered skill assessment question selected from the current answer state, or null when complete."
                    },
                    "400": {
                        "description": "Validation error in skill assessment answers."
                    },
                    "404": {
                        "description": "Assessment session was not found."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AnswerSubmit": {
                "type": "object",
                "properties": {
                    "question_id": {
                        "type": "integer"
                    },
                    "option_id": {
                        "type": "integer"
                    },
                    "scale_value": {
                        "type": "integer"
                    },
                    "response_time_ms": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "confidence_indicator": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/AnswerSubmitConfidenceIndicatorEnum"
                            },
                            {
                                "$ref": "#/components/schemas/BlankEnum"
                            }
                        ]
                    }
                },
                "required": [
                    "question_id"
                ]
            },
...
            "AssessmentHistory": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "phase": {
                        "$ref": "#/components/schemas/PhaseEnum"
                    },
                    "status": {
                        "$ref": "#/components/schemas/StatusEnum"
                    },
                    "answers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AnswerHistory"
                        },
                        "readOnly": true
                    }
                },
                "required": [
                    "answers",
                    "id"
                ]
            },
...
            "SkillAssessmentNextQuestionRequest": {
                "type": "object",
                "properties": {
                    "answers": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "integer",
                            "maximum": 5,
                            "minimum": 1
                        }
                    }
                }
            },
            "SkillAssessmentNextQuestionResponse": {
                "type": "object",
                "properties": {
                    "next_question": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/SkillAssessmentQuestion"
                            }
                        ],
                        "nullable": true
                    }
                },
                "required": [
                    "next_question"
                ]
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            },
            "tokenAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Token-based authentication with required prefix \"Token\""
            }
        }
    }
}